Completed login style and login functions, also redirect on successful login/signup to dash.html
