--- Authentication/index.html
@@ -1,44 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <!-- Link my stylesheet -->
-    <link rel='stylesheet' href='./assets/css/style.css'>
-   <title>Manhattan Crypts</title>
-</head>
-<body>
-
-    <!-- Button to open the modal login form -->
-    <!-- <button onclick="document.getElementById('loginModal').style.display='block'">Login</button> -->
-
-    <!-- The Modal -->
-    <div id="loginModal" class="modal">
-    <span onclick="document.getElementById('loginModal').style.display='none'"
-    class="closeModal" title="Close Modal">&times;</span>
-
-    <!-- Modal Content -->
-    <form class="modal-content animate">
-        <div class="imgcontainer">
-            <h2>Manhattan Crypts</h2>
-            <img src="./assets/img/avatar.png" alt="Avatar" class="avatar"></div>
-
-        <div class="container">
-            <label for="email"><b>Email:</b></label>
-            <input type="email" placeholder="Email Address" id="email"  name="email" required>
-
-            <label for="password"><b>Password:</b></label>
-            <input type="password" placeholder="Password" id="passsword" name="passsword" required>
-
-            <button type="submit" id="signIn">Sign In</button>
-            <button type="submit" id="signUp">Sign Up</button></div></div>
-    </form></div></div>
-
-    <!-- Link Firebase & my JScript -->
-    <script src="https://www.gstatic.com/firebasejs/9.5.0/firebase-app.js"></script>
-    <script src="https://www.gstatic.com/firebasejs/9.5.0/firebase-auth.js"></script>
-    <script type="module" src="./assets/js/index.js"></script>
-
-</body>
-</html>
