+++ index.html
@@ -1,0 +1,38 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <!-- Link my stylesheet -->
+    <link rel='stylesheet' href="./assets/css/style.css">
+    <link href='https://fonts.googleapis.com/css?family=Ubuntu:500' rel='stylesheet' type='text/css'>
+
+   <title>Manhattan Crypts</title>
+</head>
+<body>
+
+  <div class="login">
+    <div class="login-header">
+    </div>
+    <div class="login-form">
+      <h1>Login</h1>
+      <h3>Email:</h3>
+      <input type="email" placeholder="Email" id="email" required/><br>
+      <h3>Password:</h3>
+      <input type="password" placeholder="Password" id="pass" required/>
+      <br>
+      <input type="button" value="Login" id="signin" class="login-button"/>
+      <br>
+      <a class="sign-up" id="signup">Sign Up!</a>
+
+    </div>
+  </div>
+
+    <!-- Link Firebase & my JScript -->
+    <script src="https://www.gstatic.com/firebasejs/9.5.0/firebase-app.js"></script>
+    <script src="https://www.gstatic.com/firebasejs/9.5.0/firebase-auth.js"></script>
+    <script type="module" src="./assets/js/index.js"></script>
+
+</body>
+</html>
